// src/models/auth.ts
export interface User {
    id: string;
    name: string;
    email: string;
    role: "teacher" | "student";
    createdAt: string;
    updatedAt?: string;
}

export interface LoginRequest {
    email: string;
    password: string;
}

export interface LoginResponse {
    token: string;
    user: User;
}

export interface RegisterRequest {
    name: string;
    email: string;
    password: string;
}

export interface RegisterResponse {
    token: string;
    user: User;
}